pyproject.toml
[tool.poetry]
name = "quick_drf"
version = "0.1.0"
description = ""
authors = ["sobadgirl <drcbin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
drf-extensions = {git = "https://github.com/sobadgirl/drf-extensions.git", rev = "master"}
django-debug-toolbar = "^3.2.4"
rest-framework-generic-relations = "^2.1.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
django-cors-headers = "^3.11.0"
django-environ = "^0.8.1"
django-extensions = "^3.1.5"
django-restql = "^0.15.2"
drf-spectacular = "^0.22.0"

[tool.poetry.dev-dependencies]
prompt-toolkit = "3.0.28"
ipython = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
